feat: schemas are ordered based on dependencies
previously we output schemas in alphabetically sorted order, same as we
do for types.

this is problematic as the schemas are runtime code, and if the order is
incorrect this causes "use before declaration" type errors.

after this change we instead group the scheams based on their
dependencies and sort each group such that depended on schemas are
output before the schemas that depend on them.

the implementation isn't perfect - there are a few edge cases that it
doesn't cover, but it's a big improvement on the previous behavior.

as a bonus, also adjusted things such that more scheams will be
extracted to the schemas.ts file rather than inlined, making them
reusuable and the code easier to read.

this will fix the build problems on https://github.com/mnahkies/openapi-code-generator/pull/31
as it was failing due to the ordering issue