fix: switch to qs and add tests for query string
- switch to qs package as this advertises support for arrays/objects
  more so than the built-in node:querystring library

- add tests for query string serialization

I'm not 100% sure that this is actually what we want, or at least what
we always want. It could well be that we need to provide configuration
to control how arrays/objects are handled for serialization, as I
suspect different API's will have different opinions on this.

Currently it's also ignoring the serialization hint parameters like
style/explode - just adding a todo for now.