fix: angular query param handling sorta works
HttpParams is immutable, and so previously the query parameter handling
didn't work at all.

Now it has an expanded type to support the stripe client, and also
correctly returns the final object.

However, it does not yet support array or objects passed as a query
parameter and will throw an exception if this happens.