fix: don't coerce strings
coercing strings causes issues with objects like `{foo: undefined}`
as it coerces `undefined` to `"undefined"` rather than leaving it as
`undefined` / throwing an error if the property is not optional.

coercing numbers doesn't seem to suffer the same issue, and is needed
for parameters from route params / query strings since these always
start as a string.

the schema builder should probably become aware of when to
expect to receive a parsed JSON value rather than a string and adjust
it's use of coerce accordingly, but this will improve things in the
meantime.