feat(typescript-koa): introduce optional "respond" pattern
introduces a new pattern for responding to requests, the respond parameter.

this is essentially a strongly typed response builder that can be used to
build the return type instead of returning a plain object.

the intention is to improve the developer experience with a method chaining
API that reduces the key strokes to sensible auto completion, and avoids the
need for `as const` everywhere for type inference to work.

it's optional and completely backwards compatible for now, but as additional
features like response headers are implemented this may change.

see integration-tests/typescript-koa/src/petstore-expanded.yaml.ts for a
concrete example of before / after.