feat: add typescript-axios template (#83)
- new template `typescript-axios` based on https://github.com/axios/axios
- compared to the `typescript-fetch` template it's quite "lossy", eg:
you don't retain as much information about possible error responses. as
we can't type hint possible exceptions I don't see a way to avoid this
(short of catching and returning a result type, but this would seem to
defeat the purpose of using `axios`)
- included a couple of scripts using the github api as a test / sample

Some dog fooding has been done, but not as much as the fetch template
yet. Likely to some rough edges that will need to be fixed forward.