feat: use pre-compiled ajv validators at runtime
this builds on the speed up from memorizing the validator in #140,
further improving the speed of the zod spec file from `~1s` -> `~200ms`
(a lot better than the original `~6s`!)

there's no longer any need to memorize the validator, since the require
cache effectively does this for us now.

the compiled validation function is pretty large, but committing it
will keep me honest and prove reproducibility thanks to the CI check
for uncommitted changes after running the build / tests.