fix!: stricter booleans (#146)
Booleans now must be one of:
- `true | "true" | 1`  -> `true`
- `false | "false" | 0` -> `false`
- anything else -> error

Where previously any truthy value would be accepted as `true`

Introduces the concept of a "static schema" to create a
`PermissiveBoolean` schema to aid with generation. This is a little
annoying to unit test as it introduces `import` statements that don't
work in the test harness properly.