feat: add e2e test suite (#235)
Until now, whilst some of the unit tests execute specific pieces of
generated code using the `node:vm` module, and the integration tests
check that typescript is happy to build the code, we haven't had very
comprehensive end-to-end tests in the repo.

This has meant manually checking changes in other (mostly private)
projects, which is a bit of a pain and not very transparent.

This PR aims to add end to end tests that actually run a koa service,
and make calls to it using a generated client. Coverage will be
relatively low to begin with and iteratively improved on over time.

It will hopefully also serve as a decent sample of what a service using
this project might look like.