fix: support exactOptionalPropertyTypes on client SDKs (#278)
previously we were always treating it as `false`, rather than using the
value we read from the `tsconfig.json` file. this was intentional at the
time in #131
> - For `typescript-koa` include `| undefined` on optional properties
when `exactOptionalPropertyTypes` is configured
> - This shouldn't be required for the clients, as in this case the
consumer can choose to omit keys
>     - Though maybe it's more ergonomic to follow this behavior anyway?

However it turns out this can cause issues with the `zod` schemas when
runtime validation is enabled, and so it's better to follow this `|
undefined` pattern on the clients as well. It's probably more ergonomic
as well tbh.