fix(client-sdks): handle 204 empty responses with runtime validation (#288)
Adds e2e tests for endpoints returning `204` / empty responses and
relaxes runtime schema validation for clients to use `any`, allowing for
`axios` returning an `""`

Keeps the type as `void` which is somewhat a lie, but I believe a
pragmatic decision - in practice you shouldn't be looking at the
response body for a `204` at all, so it shouldn't really matter. The
`fetch` client will continue to throw a JSON parse error if you attempt
to read the empty response, which I think is fine for the same reason.

An alternative might've been to use a `.transform(it => undefined)` to
coerce the runtime value to `undefined` but I don't think there is all
that much value in this.

closes #285