feat: add an interactive playground to the documentation site (#216)
Oh boy was this more of a pain than I expected. 

Creates an interactive playground similar to the typespec or typescript
ones, based on https://github.com/microsoft/monaco-editor

The actual code generation is off-loaded to a web worker, which actually
made the react side of things much simpler and easier to reason about,
and reduces blocking on the UI thread/main event loop.

Prettier is used instead of Biome due to difficulties getting the WASM
to load correctly in the browser / with nextjs. Unfortunately Prettier
hits a maximum recursion limit when formatting some of the larger files
(in firefox at least), so it would be good to revisit this and try and
get Biome working.

When loading the larger files, the performance of the editor gets a bit
shaky. I'm unsure if there is a way to offload more of the editor to web
workers to reduce strain on the UI thread.

**Demo Video**

https://github.com/user-attachments/assets/5b34f63e-9eaf-4e27-88e9-5a6a0bf5b9fe