feat!: making enum evolution safer (#291)
It's useful to specify `enum` on fields in terms of documentation and
nudging users to handle the full set of values correctly. However it
does come with a pretty significant downside when thinking about clients
where you may not have full control of the release cycle (eg: native
mobile apps where you're at the mercy of the users as to when they
update).

Currently the `enum` schema validation is very strict and will throw an
error on unrecognized values, which then makes it difficult to safely
extend the schema to include new values.

This change aims to find a middle-ground where we still get the benefits
of `enum` for our client SDK's, but we also force consumers to
anticipate new values being added and handle this case.

It's configurable behind the `--enum-extensibility <value>` flag, and
additionally configurable for each individual enum value using a
`x-enum-extensibility: <value>` extension, where value can be `open` or
`closed`.

By default server enum handling remains strict/closed, and clients will
use open enums.

Potentially breaking change.