feat: typescript-express server template (#324)
introduces a second server template, `typescript-express`

**Tasks**
- [x] Template
- [x] Runtime
- [x] Integration tests
- [x] Existing E2E tests
- [x] Additional E2E tests
- [x] Documentation
- [x] Review

**Why Express**
`express` is one of the most popular server frameworks for NodeJS,
getting approximately 10x as many weekly downloads as `koa`

Its also quite similar to `koa` making it a good candidate for the
second server template.

This makes it somewhat less interesting to implement, compared to other
options like `nextjs` (#152) but also means that its a good way to
prompt refactors like #316 without requiring a bunch of new
functionality.

**Runtime**
It's clear at this point that there is a lot of duplicated code between
all the runtimes. I like keeping the separate runtime packages for
several reasons:
- Dependency declaration / management is cleaner
- NPM download trends can function as a rough proxy for adoption / usage
_(although private caching registries cause this to under-report
significantly)_

It probably makes sense to introduce a `typescript-runtime-common`
package soon.

**Approach**
The approach ends up looking near identical to the `typescript-koa`
template, in terms of the end developer experience. This is particularly
reinforced by the E2E tests and how little difference exists between the
two implementations.