chore: switch to pnpm and fix peer dependencies (#345)
A somewhat reoccurring issue has been dependency updates introducing
`peerDependency` misalignment that causes installation / run errors with
tools like `npx`.

Originally I was planning to address this by introducing new CI steps
that checked for this, but instead I've decided to give `pnpm` a go. As
far as I understand it has the strictest rules of the common package
managers around dependency resolution, and therefore if it works with
`pnpm` it'll probably work with everything else.

Making this change surfaced a number of implicit dependency issues
across the repository, and unintentional reliance on `jest` globals.

I'd be fairly keen to additionally do away with the `lerna` dependency,
as its now only used for publishing the packages, but that's probably a
job for another day.