feat: boolean schemas support enum (#351)
adds support for the `enum` keyword to `boolean` schemas, which is handy
for creating discriminated unions on occasion.

after this change a schema like:
```
type: boolean
enum:
  - true
```
will only allow `true` to be passed, vice-versa for `false`.

I've opted to not include the normal enum open/closed concept here as it
seems a bit unnecessary when the enums domain can only contain two
values at most - If open enums are expected to be needed then you should
probably be using a `string` or `number` enum.

requested by @markios