feat: support application/x-www-form-urlencoded request bodies (#352)
adds support for `application/x-www-form-urlencoded` request bodies,
including the `encoding` property to customize the serialization of
object and array values.

it's implemented for `typescript-fetch` and `typescript-axios` with
partial support on the server templates (they don't currently understand
encoding, so its a bit of pot-luck whether they'll accurately
de-serialize the data)

later we'll extract some of the duplicated code into a new core runtime
package, as there's an increasingly large body of code that is shared
between the different runtimes.

additionally improves handling of optional request bodies, particularly
in the case that the consuming project has `exactOptionalPropertyTypes`
enabled. This is currently more correct on the `typescript-fetch`
template, due to some quirks of `axios` - see skipped `e2e` tests for
details.