feat!: move virtual schemas (#398)
this is an annoying but necessary change that is blocking a number of
improvements.

in essence we are moving the responsibility for generating "virtual
schemas", ie: giving a name to otherwise anonymous entities into the
`Input` class, and making the `OpenapiLoader` instance private, as it
always should've been.

this is a step towards more clearly delimiting between the internal
domain model and the external (eg: OAS) domain model, which is going to
unlock improvements to support for 3.1 / 3.2 functionality.

unfortunately, this also identified that there were some inconsistencies
in the names being generated, and despite my best efforts I wasn't able
to come up with a clean way to preserve the original names.

given that it was going to be a breaking change anyway, I've
doubled-down a bit and allowed myself to further change the names,
notably omitting the media type if only a single media type is defined
by the specification in that context (eg: there is no ambiguity by
omitting it)

this change also "unwraps" some types that were adding needless
indirection, eg:
```
type SomeModel = {...}
type SomeRequestBody = SomeModel
```
will now just directly reference `SomeModel` where needed.

whilst not yet exposed through the CLI, it also internally introduces a
`SyntheticNameGenerator` interface that will later be exposed in some
way to allow users to customize the names generated going forward.

BREAKING CHANGE: generated names of some output types/schemas have
changed, sorry