fix!: query param serialization/parsing (#381)
respect the `style` and `explode` modifiers on `in: query` parameters.

this is required to generate working clients for
https://github.com/typesense/typesense-api-spec for example, as it
defines the query parameters as a single object that should be
"exploded" (eg: flattened) to multiple parameters.

removes the `qs` dependency in favor of our own parsing / serialization
based on the `style` + `explode` modifiers.

in addition to the e2e tests in this repository, I've tested the
generated clients e2e against a typesense server, which previously
didn't work.